fix 1

+++ greenplumpython/core/gpapply.py
@@ -28,12 +28,18 @@ def pythonExec(df, funcName, typeName, output, extra_args):
     return select_func
 
 def gpApply(dataframe, py_func, db, output, clear_existing = True, runtime_id = 'plc_python', runtime_type = 'plpythonu', **kwargs):
+    if py_func == None:
+        raise ValueError("No input function provided")
+    
     s = inspect.getsource(py_func)
-    #gpdb_tbl_name = "testtbl"
     function_name = randomString()
     typeName = randomStringType()
     params = []
-    columns = []
+    columns = inspect.getfullargspec(py_func)[0]
+
+    if dataframe == None:
+        raise ValueError("No input dataframe provided")
+
     for i, col in enumerate(dataframe.table_metadata.signature):
         for j, column in enumerate(col):
             params.append(column+" "+col[column])
@@ -45,6 +51,9 @@ def gpApply(dataframe, py_func, db, output, clear_existing = True, runtime_id = 
     for key, value in kwargs.items(): 
         params.append(columns[args_index] + " " + str(value[1]))
         args_index = args_index + 1
+    
+    if output == None or output.signature == None:
+        raise ValueError("Output.signature must be provided")
         
     create_type_sql = createTypeFunc(output.signature, typeName)
     runtime_id_str = ''
@@ -53,15 +62,24 @@ def gpApply(dataframe, py_func, db, output, clear_existing = True, runtime_id = 
     function_body = "CREATE OR REPLACE FUNCTION %s(%s) RETURNS %s AS $$\n%s\n%s\nreturn %s(%s) $$ LANGUAGE %s;" % (function_name,",".join(params),typeName,runtime_id_str,s,py_func.__name__,",".join(columns),runtime_type)
     select_sql = pythonExec(dataframe, function_name, typeName, output, kwargs)
     drop_sql = "DROP TYPE " + typeName + " CASCADE;"
+
+    if db == None:
+        raise ValueError("No database connection provided")
+
     if clear_existing:
         drop_table_sql = "drop table if exists %s;" % output.name
         db.execute(drop_table_sql)
     db.execute(create_type_sql)
-    db.execute(function_body)
-    res = None
-    if output.name == None or output.name == "":
-        res = db.execute_query(select_sql)
-    else:
-        db.execute(select_sql)
+    try:
+        db.execute(function_body)
+        res = None
+        if output.name == None or output.name == "":
+            res = db.execute_query(select_sql)
+        else:
+            db.execute(select_sql)
+    except Exception as e:
+        db.execute(drop_sql)
+        raise e
+
     db.execute(drop_sql)
     return res
